can: implement slcan and hlcan switch Added the ability to switch between slcan and hlcan drivers via sysfs to avoid conflicts between them. hlcan is based on slcan and is used for the Chinese CAN analyzer QinHeng Electronics HL-340. It is derived from the kobolt/usb-can package and the Linux slcan driver.
The HL-340 is essentially a USB-to-serial converter, for which Linux provides the ch341-uart driver and creates a /dev/ttyUSB device.
This driver implements the necessary part of the serial protocol to communicate with that hardware.

Removed as SUbmodule and cloned the hlcan repo: https://github.com/V0lk3n/usb-can-2-module

git cherry-pick 326239c85829daf431ce19adf7607d585d2d4d82